+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -70,7 +70,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01ke6fm0yg2dm26zz8a7nr60jp"
+        "conv_01ke92zwsmncqtk5zwq212psrj"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """