+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-05T07:05:41.258Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-06T07:22:42.672Z")
     r"""The date and time the resource was last updated"""