+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KE6FM1C2SJ71A7TN5K9F5AX5"
+        "01KE92ZX685CYZE6J064TEDKCW"
     )
     r"""The unique identifier of the data source"""