+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KE6FM1C5YNBS6GCKRQH5TYNV"
+        "01KE92ZX6HR3R5K0B4Z2W0J6VY"
     )
     r"""The unique identifier of the data source"""