+++ packages/orq-rc/src/orq_ai_sdk/router.py
@@ -4,6 +4,10 @@ from .basesdk import BaseSDK
 from enum import Enum
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
+from orq_ai_sdk.models import (
+    createchatcompletionop as models_createchatcompletionop,
+    createimageop as models_createimageop,
+)
 from orq_ai_sdk.types import OptionalNullable, UNSET
 from orq_ai_sdk.utils import eventstreaming, get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -20,15 +24,15 @@ class Router(BaseSDK):
         self,
         *,
         messages: Union[
-            List[models.CreateChatCompletionMessages],
-            List[models.CreateChatCompletionMessagesTypedDict],
+            List[models_createchatcompletionop.CreateChatCompletionMessages],
+            List[models_createchatcompletionop.CreateChatCompletionMessagesTypedDict],
         ],
         model: str,
         metadata: Optional[Dict[str, str]] = None,
         audio: OptionalNullable[
             Union[
-                models.CreateChatCompletionAudio,
-                models.CreateChatCompletionAudioTypedDict,
+                models_createchatcompletionop.CreateChatCompletionAudio,
+                models_createchatcompletionop.CreateChatCompletionAudioTypedDict,
             ]
         ] = UNSET,
         frequency_penalty: OptionalNullable[float] = UNSET,
@@ -40,29 +44,31 @@ class Router(BaseSDK):
         presence_penalty: OptionalNullable[float] = UNSET,
         response_format: Optional[
             Union[
-                models.CreateChatCompletionResponseFormat,
-                models.CreateChatCompletionResponseFormatTypedDict,
+                models_createchatcompletionop.CreateChatCompletionResponseFormat,
+                models_createchatcompletionop.CreateChatCompletionResponseFormatTypedDict,
             ]
         ] = None,
-        reasoning_effort: Optional[models.CreateChatCompletionReasoningEffort] = None,
+        reasoning_effort: Optional[
+            models_createchatcompletionop.CreateChatCompletionReasoningEffort
+        ] = None,
         verbosity: Optional[str] = None,
         seed: OptionalNullable[float] = UNSET,
         stop: OptionalNullable[
             Union[
-                models.CreateChatCompletionStop,
-                models.CreateChatCompletionStopTypedDict,
+                models_createchatcompletionop.CreateChatCompletionStop,
+                models_createchatcompletionop.CreateChatCompletionStopTypedDict,
             ]
         ] = UNSET,
         stream_options: OptionalNullable[
             Union[
-                models.CreateChatCompletionStreamOptions,
-                models.CreateChatCompletionStreamOptionsTypedDict,
+                models_createchatcompletionop.CreateChatCompletionStreamOptions,
+                models_createchatcompletionop.CreateChatCompletionStreamOptionsTypedDict,
             ]
         ] = UNSET,
         thinking: Optional[
             Union[
-                models.CreateChatCompletionThinking,
-                models.CreateChatCompletionThinkingTypedDict,
+                models_createchatcompletionop.CreateChatCompletionThinking,
+                models_createchatcompletionop.CreateChatCompletionThinkingTypedDict,
             ]
         ] = None,
         temperature: OptionalNullable[float] = UNSET,
@@ -70,29 +76,32 @@ class Router(BaseSDK):
         top_k: OptionalNullable[float] = UNSET,
         tools: Optional[
             Union[
-                List[models.CreateChatCompletionTools],
-                List[models.CreateChatCompletionToolsTypedDict],
+                List[models_createchatcompletionop.CreateChatCompletionTools],
+                List[models_createchatcompletionop.CreateChatCompletionToolsTypedDict],
             ]
         ] = None,
         tool_choice: Optional[
             Union[
-                models.CreateChatCompletionToolChoice,
-                models.CreateChatCompletionToolChoiceTypedDict,
+                models_createchatcompletionop.CreateChatCompletionToolChoice,
+                models_createchatcompletionop.CreateChatCompletionToolChoiceTypedDict,
             ]
         ] = None,
         parallel_tool_calls: Optional[bool] = None,
         modalities: OptionalNullable[
-            List[models.CreateChatCompletionModalities]
+            List[models_createchatcompletionop.CreateChatCompletionModalities]
         ] = UNSET,
         guardrails: Optional[
             Union[
-                List[models.CreateChatCompletionGuardrails],
-                List[models.CreateChatCompletionGuardrailsTypedDict],
+                List[models_createchatcompletionop.CreateChatCompletionGuardrails],
+                List[
+                    models_createchatcompletionop.CreateChatCompletionGuardrailsTypedDict
+                ],
             ]
         ] = None,
         orq: Optional[
             Union[
-                models.CreateChatCompletionOrq, models.CreateChatCompletionOrqTypedDict
+                models_createchatcompletionop.CreateChatCompletionOrq,
+                models_createchatcompletionop.CreateChatCompletionOrqTypedDict,
             ]
         ] = None,
         stream: Optional[bool] = False,
@@ -284,15 +293,15 @@ class Router(BaseSDK):
         self,
         *,
         messages: Union[
-            List[models.CreateChatCompletionMessages],
-            List[models.CreateChatCompletionMessagesTypedDict],
+            List[models_createchatcompletionop.CreateChatCompletionMessages],
+            List[models_createchatcompletionop.CreateChatCompletionMessagesTypedDict],
         ],
         model: str,
         metadata: Optional[Dict[str, str]] = None,
         audio: OptionalNullable[
             Union[
-                models.CreateChatCompletionAudio,
-                models.CreateChatCompletionAudioTypedDict,
+                models_createchatcompletionop.CreateChatCompletionAudio,
+                models_createchatcompletionop.CreateChatCompletionAudioTypedDict,
             ]
         ] = UNSET,
         frequency_penalty: OptionalNullable[float] = UNSET,
@@ -304,29 +313,31 @@ class Router(BaseSDK):
         presence_penalty: OptionalNullable[float] = UNSET,
         response_format: Optional[
             Union[
-                models.CreateChatCompletionResponseFormat,
-                models.CreateChatCompletionResponseFormatTypedDict,
+                models_createchatcompletionop.CreateChatCompletionResponseFormat,
+                models_createchatcompletionop.CreateChatCompletionResponseFormatTypedDict,
             ]
         ] = None,
-        reasoning_effort: Optional[models.CreateChatCompletionReasoningEffort] = None,
+        reasoning_effort: Optional[
+            models_createchatcompletionop.CreateChatCompletionReasoningEffort
+        ] = None,
         verbosity: Optional[str] = None,
         seed: OptionalNullable[float] = UNSET,
         stop: OptionalNullable[
             Union[
-                models.CreateChatCompletionStop,
-                models.CreateChatCompletionStopTypedDict,
+                models_createchatcompletionop.CreateChatCompletionStop,
+                models_createchatcompletionop.CreateChatCompletionStopTypedDict,
             ]
         ] = UNSET,
         stream_options: OptionalNullable[
             Union[
-                models.CreateChatCompletionStreamOptions,
-                models.CreateChatCompletionStreamOptionsTypedDict,
+                models_createchatcompletionop.CreateChatCompletionStreamOptions,
+                models_createchatcompletionop.CreateChatCompletionStreamOptionsTypedDict,
             ]
         ] = UNSET,
         thinking: Optional[
             Union[
-                models.CreateChatCompletionThinking,
-                models.CreateChatCompletionThinkingTypedDict,
+                models_createchatcompletionop.CreateChatCompletionThinking,
+                models_createchatcompletionop.CreateChatCompletionThinkingTypedDict,
             ]
         ] = None,
         temperature: OptionalNullable[float] = UNSET,
@@ -334,29 +345,32 @@ class Router(BaseSDK):
         top_k: OptionalNullable[float] = UNSET,
         tools: Optional[
             Union[
-                List[models.CreateChatCompletionTools],
-                List[models.CreateChatCompletionToolsTypedDict],
+                List[models_createchatcompletionop.CreateChatCompletionTools],
+                List[models_createchatcompletionop.CreateChatCompletionToolsTypedDict],
             ]
         ] = None,
         tool_choice: Optional[
             Union[
-                models.CreateChatCompletionToolChoice,
-                models.CreateChatCompletionToolChoiceTypedDict,
+                models_createchatcompletionop.CreateChatCompletionToolChoice,
+                models_createchatcompletionop.CreateChatCompletionToolChoiceTypedDict,
             ]
         ] = None,
         parallel_tool_calls: Optional[bool] = None,
         modalities: OptionalNullable[
-            List[models.CreateChatCompletionModalities]
+            List[models_createchatcompletionop.CreateChatCompletionModalities]
         ] = UNSET,
         guardrails: Optional[
             Union[
-                List[models.CreateChatCompletionGuardrails],
-                List[models.CreateChatCompletionGuardrailsTypedDict],
+                List[models_createchatcompletionop.CreateChatCompletionGuardrails],
+                List[
+                    models_createchatcompletionop.CreateChatCompletionGuardrailsTypedDict
+                ],
             ]
         ] = None,
         orq: Optional[
             Union[
-                models.CreateChatCompletionOrq, models.CreateChatCompletionOrqTypedDict
+                models_createchatcompletionop.CreateChatCompletionOrq,
+                models_createchatcompletionop.CreateChatCompletionOrqTypedDict,
             ]
         ] = None,
         stream: Optional[bool] = False,
@@ -549,17 +563,22 @@ class Router(BaseSDK):
         *,
         prompt: str,
         model: str,
-        background: OptionalNullable[models.Background] = UNSET,
-        moderation: OptionalNullable[models.Moderation] = UNSET,
+        background: OptionalNullable[models_createimageop.Background] = UNSET,
+        moderation: OptionalNullable[models_createimageop.Moderation] = UNSET,
         n: OptionalNullable[int] = 1,
         output_compression: OptionalNullable[int] = UNSET,
-        output_format: OptionalNullable[models.OutputFormat] = UNSET,
-        quality: OptionalNullable[models.Quality] = UNSET,
-        response_format: OptionalNullable[models.CreateImageResponseFormat] = UNSET,
+        output_format: OptionalNullable[models_createimageop.OutputFormat] = UNSET,
+        quality: OptionalNullable[models_createimageop.Quality] = UNSET,
+        response_format: OptionalNullable[
+            models_createimageop.CreateImageResponseFormat
+        ] = UNSET,
         size: OptionalNullable[str] = UNSET,
-        style: OptionalNullable[models.Style] = UNSET,
+        style: OptionalNullable[models_createimageop.Style] = UNSET,
         orq: Optional[
-            Union[models.CreateImageOrq, models.CreateImageOrqTypedDict]
+            Union[
+                models_createimageop.CreateImageOrq,
+                models_createimageop.CreateImageOrqTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -674,17 +693,22 @@ class Router(BaseSDK):
         *,
         prompt: str,
         model: str,
-        background: OptionalNullable[models.Background] = UNSET,
-        moderation: OptionalNullable[models.Moderation] = UNSET,
+        background: OptionalNullable[models_createimageop.Background] = UNSET,
+        moderation: OptionalNullable[models_createimageop.Moderation] = UNSET,
         n: OptionalNullable[int] = 1,
         output_compression: OptionalNullable[int] = UNSET,
-        output_format: OptionalNullable[models.OutputFormat] = UNSET,
-        quality: OptionalNullable[models.Quality] = UNSET,
-        response_format: OptionalNullable[models.CreateImageResponseFormat] = UNSET,
+        output_format: OptionalNullable[models_createimageop.OutputFormat] = UNSET,
+        quality: OptionalNullable[models_createimageop.Quality] = UNSET,
+        response_format: OptionalNullable[
+            models_createimageop.CreateImageResponseFormat
+        ] = UNSET,
         size: OptionalNullable[str] = UNSET,
-        style: OptionalNullable[models.Style] = UNSET,
+        style: OptionalNullable[models_createimageop.Style] = UNSET,
         orq: Optional[
-            Union[models.CreateImageOrq, models.CreateImageOrqTypedDict]
+            Union[
+                models_createimageop.CreateImageOrq,
+                models_createimageop.CreateImageOrqTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,